streaming/consumer/requirements.txt
kafka-python==2.0.2
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.2
